refactor(fe): apply Modal for ProblemsUploadButton
